fix: ensure float32 type for tw in run_moe_torch function in collect_moe_v1.py to comply with fused_marlin_moe requirements

+++ collector/vllm/collect_moe_v1.py
@@ -538,7 +538,7 @@ def run_moe_torch(
                 w1_int4_scale,
                 w2_int4_scale,
                 None,  # gating_output (unused when topk_weights/ids provided)
-                tw,
+                tw.float(),  # fused_marlin_moe asserts float32
                 ti,
                 _int4_wo_quant_type_id,
                 w1_zeros=None,
